+++ src/Dto/Customer/Customer/Get.php
@@ -27,6 +27,7 @@ use Shopgate\ConnectSdk\Dto\Customer\Customer\Dto\Contact;
 use Shopgate\ConnectSdk\Dto\Customer\Customer\Dto\Settings;
 
 /**
+ * @method string getId()
  * @method string getExternalCustomerNumber()
  * @method string getFirstName()
  * @method string getMiddleName()
@@ -34,6 +35,7 @@ use Shopgate\ConnectSdk\Dto\Customer\Customer\Dto\Settings;
  * @method string getEmailAddress()
  * @method string getStatus()
  * @method boolean getIsAnonymous()
+ * @method string getExternalUpdateDate()
  * @method Contact[] getContacts()
  * @method Attribute[] getAttributes()
  * @method Settings getSettings()